refactor : OpenStatus 클래스 이동 - 테스트 import 수정
